WIP: move sql-composer serde functionality into its own crate
