+++ sql-composer-cli/Cargo.toml
@@ -7,22 +7,24 @@ edition = "2018"
 [[bin]]
 name = "sqlc"
 path = "src/main.rs"
-required-features = ["sql-composer/composer-serde"]
 
 [features]
-dbd-mysql = ["mysql"]
-dbd-rusqlite = ["rusqlite"]
-dbd-postgres = ["postgres"]
+dbd-mysql = ["mysql", "sql-composer-mysql/composer-serde"]
+dbd-rusqlite = ["rusqlite", "sql-composer-rusqlite/composer-serde"]
+dbd-postgres = ["postgres", "sql-composer-postgres/composer-serde"]
 
 [dependencies]
 quicli = "0.4.0"
-sql-composer={ path="../sql-composer" }
+sql-composer={ path="../sql-composer", version = "0.1.0" }
+sql-composer-mysql={ path="../sql-composer-mysql", optional = true }
+sql-composer-rusqlite={ path="../sql-composer-rusqlite", optional = true }
+sql-composer-postgres={ path="../sql-composer-postgres", optional = true }
 structopt = "0.2.15"
 mysql = { version = "16.0.2", optional = true }
 rusqlite = { version = "0.17.0", optional = true }
 postgres = { version = "0.15.2", optional = true }
 serde = "1.0.89"
 serde_json = "1.0.39"
-serde_postgres = "0.1.3"
+serde_postgres = { version = "0.1.3", optional = true }
 serde-transcode = "1.1.0"
 serde-value = "0.6.0"
