+++ sql-composer-mysql/Cargo.toml
@@ -5,13 +5,10 @@ authors = ["Kenny Flegal <soulstompp@gmail.com>"]
 edition = '2018'
 
 [features]
-composer-serde = ["serde", "serde-value"]
+composer-serde = ["serde-value"]
 
 [dependencies]
-#chrono = "0.4"
 dotenv = "0.15.0"
 mysql  = "16.0.2"
 sql-composer={ path="../sql-composer" }
-serde = { version = "1.0.89", optional = true }
 serde-value = { version = "0.6.0", optional = true }
-#time = "0.1.40"
