+++ sql-composer-serde/Cargo.toml
@@ -1,15 +1,11 @@
 [package]
-name = "sql-composer"
+name = "sql-composer-serde"
 version = "0.1.0"
 authors = ["Kenny Flegal <soulstompp@gmail.com>"]
 edition = '2018'
 
-[features]
-composer-serde = []
-
 [dependencies]
-chrono = "0.4"
-dotenv = "0.15.0"
-error-chain = "0.12.1"
 nom = "5.0.1"
 nom_locate = "1.0.0"
+sql-composer = { path="../sql-composer" }
+serde-value = { version = "0.6.0"}
